Add Type Ahead
A input form that allows a instant (on the fly) search of cities and states. Working with fetch() and promises as well as Regular Expressions. The data is fetched from an external endpoint and stored into an array. The array is filtered on the user input to only display matching cities or states.